backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pydantic==1.10.2
email-validator==1.3.0
python-dotenv==0.21.1
aiofiles==22.1.0
jinja2==3.1.2
pytest==7.2.0
httpx==0.23.3
SQLAlchemy-Utils==0.38.3 